+++ Dip/Dip/Dip.swift
@@ -27,9 +27,9 @@ import Foundation
 // MARK: - DependencyContainer
 
 /**
- * _Dip_'s Dependency Containers allow you to do very simple **Dependency Injection**
- * by associating `protocols` to concrete implementations
- */
+_Dip_'s Dependency Containers allow you to do very simple **Dependency Injection**
+by associating `protocols` to concrete implementations
+*/
 public class DependencyContainer {
   
   /**
@@ -42,8 +42,7 @@ public class DependencyContainer {
     case Int(IntegerLiteralType)
   }
   
-  private var dependencies = [DefinitionKey : Definition]()
-  private var lock: OSSpinLock = OS_SPINLOCK_INIT
+  var definitions = [DefinitionKey : Definition]()
   
   /**
    Designated initializer for a DependencyContainer
@@ -66,9 +65,30 @@ public class DependencyContainer {
   Clear all the previously registered dependencies on this container.
   */
   public func reset() {
-    lockAndDo {
-      dependencies.removeAll()
-    }
+    definitions.removeAll()
+  }
+  
+  /**
+   Removes previously registered definition from container.
+   
+   - parameter tag: tag used to register definition
+   - parameter definition: definition to remove
+  */
+  public func remove<T, F>(definition: DefinitionOf<T, F>) {
+    let key = DefinitionKey(protocolType: T.self, factoryType: F.self, associatedTag: definition.tag)
+    definitions[key] = nil
+  }
+  
+  /**
+   Registers new definiton in container and associate it with provided tag.
+   Will override already registered definition for the same type and factory associated with the same tag.
+   
+   - parameter tag: The arbitrary tag to associate definition with
+   - parameter definition: definition to register in container
+  */
+  public func register<T, F>(definition: DefinitionOf<T, F>) {
+    let key = DefinitionKey(protocolType: T.self, factoryType: F.self, associatedTag: definition.tag)
+    definitions[key] = definition
   }
   
   // MARK: Register dependencies
@@ -77,24 +97,41 @@ public class DependencyContainer {
   Register a Void->T factory associated with optional tag.
   
   - parameter tag: The arbitrary tag to associate this factory with when registering with that protocol. Pass `nil` to associate with any tag. Default value is `nil`.
+  - parameter scope: scope to use for this compone
   - parameter factory: The factory to register, with return type of protocol you want to register it for
+  - returns: definition created for provided type and factory
   
-  - note: You must cast the factory return type to the protocol you want to register it with (e.g `MyClass() as MyAPI`)
+  - note: You must cast the factory return type to the protocol you want to register it for.
+  Inside factory block if you need to reference container use it as `unowned` to avoid retain cycle.
+  
+  **Example**:
+  ```swift
+  container.register { ServiceImp() as Service }
+  container.register(tag: "service") { ServiceImp() as Service }
+  container.register(.ObjectGraph) { ServiceImp() as Service }
+  container.register { [unowned container] 
+    ClientImp(service: container.resolve() as Service) as Client 
+  }
+  ```
   */
-  public func register<T>(tag tag: Tag? = nil, factory: ()->T) -> DefinitionOf<T> {
-    return register(tag: tag, factory: factory, scope: .Prototype) as DefinitionOf<T>
+  public func register<T>(tag tag: Tag? = nil, _ scope: ComponentScope = .Prototype, factory: ()->T) -> DefinitionOf<T, ()->T> {
+    return registerFactory(tag: tag, scope: scope, factory: factory)
   }
   
   /**
    Register a Singleton instance associated with optional tag.
    
-   - parameter tag: The arbitrary tag to associate this instance with when registering with that protocol. `nil` to associate with any tag.
+   - parameter tag: The arbitrary tag to associate this instance with when registering with that protocol. 
+   Pass `nil` to associate with any tag.
    - parameter instance: The instance to register, with return type of protocol you want to register it for
    
    - note: You must cast the instance to the protocol you want to register it with (e.g `MyClass() as MyAPI`)
+   
+   **Deprecated**: Use `register(.Singleton){}` method instead to define singleton scope.
    */
-  public func register<T>(tag tag: Tag? = nil, @autoclosure(escaping) instance factory: ()->T) -> DefinitionOf<T> {
-    return register(tag: tag, factory: { factory() }, scope: .Singleton)
+  @available(*, deprecated, message="Use `register(.Singleton){}` method instead to define singleton scope.")
+  public func register<T>(tag tag: Tag? = nil, @autoclosure(escaping) instance factory: ()->T) -> DefinitionOf<T, ()->T> {
+    return registerFactory(tag: tag, scope: .Singleton, factory: { factory() })
   }
   
   /**
@@ -103,26 +140,25 @@ public class DependencyContainer {
    - parameter tag: The arbitrary tag to look for when resolving this protocol.
    - parameter factory: generic factory that should be used to create concrete instance of type
    - parameter scope: scope of the component. Default value is `Prototype`
-   
+   - returns: definition created for provided type and factory
+
    - note: You should not call this method directly, instead call any of other `register` methods.
            You _should_ use this method only to register dependency with more runtime arguments
            than _Dip_ supports (currently it's up to six) like in this example:
    
    ```swift
-   public func register<T, Arg1, Arg2, Arg3, ...>(tag: Tag? = nil, factory: (Arg1, Arg2, Arg3, ...) -> T) -> DefinitionOf<T> {
-     return register(tag: tag, factory: factory, scope: .Prototype) as DefinitionOf<T>
+   public func register<T, Arg1, Arg2, Arg3, ...>(tag: Tag? = nil, scope: ComponentScope = .Prototype, factory: (Arg1, Arg2, Arg3, ...) -> T) -> DefinitionOf<T, (Arg1, Arg2, Arg3, ...) -> T> {
+     return registerFactory(tag: tag, scope: scope, factory: factory) as DefinitionOf<T, (Arg1, Arg2, Arg3, ...) -> T>
    }
    ```
    
    Though before you do that you should probably review your design and try to reduce number of depnedencies.
    
    */
-  public func register<T, F>(tag tag: Tag? = nil, factory: F, scope: ComponentScope) -> DefinitionOf<T> {
+  public func registerFactory<T, F>(tag tag: Tag? = nil, scope: ComponentScope, factory: F) -> DefinitionOf<T, F> {
     let key = DefinitionKey(protocolType: T.self, factoryType: F.self, associatedTag: tag)
-    let definition = DefinitionOf<T>(factory: factory, scope: scope)
-    lockAndDo {
-      dependencies[key] = definition
-    }
+    let definition = DefinitionOf<T, F>(factory: factory, scope: scope, tag: tag)
+    definitions[key] = definition
     return definition
   }
   
@@ -131,9 +167,18 @@ public class DependencyContainer {
   /**
   Resolve a dependency. 
   
-  If no instance/factory was registered with this `tag` for this `protocol`, it will try to resolve the instance/factory associated with `nil` (no tag).
+  If no definition was registered with this `tag` for this `protocol`,
+  it will try to resolve the definition associated with `nil` (no tag).
   
   - parameter tag: The arbitrary tag to look for when resolving this protocol.
+  
+  **Example**:
+  ```swift
+  let service = container.resolve() as Service
+  let service = container.resolve(tag: "service") as Service
+  let service: Service = container.resolve()
+  ```
+  
   */
   public func resolve<T>(tag tag: Tag? = nil) -> T {
     return resolve(tag: tag) { (factory: ()->T) in factory() }
@@ -144,6 +189,7 @@ public class DependencyContainer {
    
    - parameter tag: The arbitrary tag to look for when resolving this protocol.
    - parameter builder: Generic closure that accepts generic factory and returns inctance produced by that factory
+   - returns: resolved instance of type T
    
    - note: You should not call this method directly, instead call any of other `resolve` methods. (see `RuntimeArguments.swift`).
            You _should_ use this method only to resolve dependency with more runtime arguments than _Dip_ supports
@@ -161,37 +207,75 @@ public class DependencyContainer {
   public func resolve<T, F>(tag tag: Tag? = nil, builder: F->T) -> T {
     let key = DefinitionKey(protocolType: T.self, factoryType: F.self, associatedTag: tag)
     let nilTagKey = tag.map { _ in DefinitionKey(protocolType: T.self, factoryType: F.self, associatedTag: nil) }
-    
-    var resolved: T!
-    lockAndDo { [unowned self] in
-      resolved = self._resolve(key, nilTagKey: nilTagKey, builder: builder)
+
+    guard let definition = (self.definitions[key] ?? self.definitions[nilTagKey]) as? DefinitionOf<T, F> else {
+      fatalError("No definition registered with " + (tag == nil ? "\(key)" : "\(key) or \(nilTagKey)") + ". "
+        + "Check the tag, type you try to resolve, number, order and types of runtime arguments passed to `resolve()`.")
     }
-    return resolved
+
+    let usingKey: DefinitionKey? = definition.scope == .ObjectGraph ? key : nil
+    return _resolve(tag, key: usingKey, definition: definition, builder: builder)
   }
   
   /// Actually resolve dependency
-  private func _resolve<T, F>(key: DefinitionKey, nilTagKey: DefinitionKey?, builder: F->T) -> T {
-    guard let definition = (self.dependencies[key] ?? self.dependencies[nilTagKey]) as? DefinitionOf<T> else {
-      fatalError("No instance factory registered with \(key) or \(nilTagKey)")
+  private func _resolve<T, F>(tag: Tag? = nil, key: DefinitionKey?, var definition: DefinitionOf<T, F>, builder: F->T) -> T {
+    
+    return resolvedInstances.resolve {
+      
+      if let previouslyResolved: T = resolvedInstances.previouslyResolved(key, definition: definition) {
+        return previouslyResolved
+      }
+      else {
+        let resolvedInstance = builder(definition.factory)
+        
+        //when builder calls factory it will in turn resolve sub-dependencies (if there are any)
+        //when it returns instance that we try to resolve here can be already resolved
+        //so we return it, throwing away instance created by previous call to builder
+        if let previouslyResolved: T = resolvedInstances.previouslyResolved(key, definition: definition) {
+          return previouslyResolved
+        }
+        
+        resolvedInstances.storeResolvedInstance(resolvedInstance, forKey: key)
+        definition.resolvedInstance(self, tag: tag, instance: resolvedInstance)
+        definition.resolveDependenciesBlock?(self, resolvedInstance)
+        
+        return resolvedInstance
+      }
+      
     }
     
-    if let resolvedInstance = definition.resolvedInstance {
-      return resolvedInstance
-    }
-    else {
-      let resolved = builder(definition.factory as! F)
-      definition.resolvedInstance = resolved
-      return resolved
-    }
   }
   
   // MARK: - Private
   
-  private func lockAndDo(@noescape block: Void->Void) {
-    OSSpinLockLock(&lock)
-    defer { OSSpinLockUnlock(&lock) }
-    block()
+  let resolvedInstances = ResolvedInstances()
+  
+  ///Pool to hold instances, created during call to `resolve()`. 
+  ///Before `resolve()` returns pool is drained.
+  class ResolvedInstances {
+    var resolvedInstances = [DefinitionKey: Any]()
+
+    func storeResolvedInstance<T>(instance: T, forKey key: DefinitionKey?) {
+      self.resolvedInstances[key] = instance
+    }
+    
+    func previouslyResolved<T, F>(key: DefinitionKey?, definition: DefinitionOf<T, F>) -> T? {
+      return (definition.resolvedInstance ?? self.resolvedInstances[key]) as? T
+    }
+    
+    private var depth: Int = 0
+    
+    func resolve<T>(@noescape block: ()->T) -> T {
+      depth++
+      let resolved = block()
+      depth--
+      if depth == 0 {
+        resolvedInstances.removeAll()
+      }
+      return resolved
+    }
   }
+  
 }
 
 extension DependencyContainer.Tag: IntegerLiteralConvertible {
@@ -229,8 +313,14 @@ public func ==(lhs: DependencyContainer.Tag, rhs: DependencyContainer.Tag) -> Bo
 }
 
 extension Dictionary {
-  subscript(key: Key?) -> Value! {
-    guard let key = key else { return nil }
-    return self[key]
+  subscript(key: Key?) -> Value? {
+    get {
+      guard let key = key else { return nil }
+      return self[key]
+    }
+    set {
+      guard let key = key else { return }
+      self[key] = newValue
+    }
   }
 }