+++ Dip/DipTests/DipTests.swift
@@ -1,9 +1,25 @@
 //
-//  DipTests.swift
-//  DipTests
+// Dip
 //
-//  Created by Ilya Puchka on 04.11.15.
-//  Copyright © 2015 AliSoftware. All rights reserved.
+// Copyright (c) 2015 Olivier Halligon <olivier@halligon.net>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 //
 
 import XCTest
@@ -13,17 +29,15 @@ protocol Service {
   func getServiceName() -> String
 }
 
-class ServiceImp1: Service {
+extension Service {
   func getServiceName() -> String {
-    return "ServiceImp1"
+    return "\(self.dynamicType)"
   }
 }
 
-class ServiceImp2: Service {
-  func getServiceName() -> String {
-    return "ServiceImp2"
-  }
-}
+class ServiceImp1: Service {}
+
+class ServiceImp2: Service {}
 
 class DipTests: XCTestCase {
   
@@ -84,28 +98,18 @@ class DipTests: XCTestCase {
     XCTAssertTrue(service2 is ServiceImp2)
   }
   
-  func testThatItResolvesTypeAsNewInstanceEveryTime() {
+  func testThatItCallsResolveDependenciesOnDefinition() {
     //given
-    container.register { ServiceImp1() as Service }
+    var resolveDependenciesCalled = false
+    container.register { ServiceImp1() as Service }.resolveDependencies(container) { (c, s) in
+      resolveDependenciesCalled = true
+    }
     
     //when
-    let service1 = container.resolve() as Service
-    let service2 = container.resolve() as Service
+    container.resolve() as Service
     
     //then
-    XCTAssertFalse((service1 as! ServiceImp1) === (service2 as! ServiceImp1))
-  }
-  
-  func testThatItReusesInstanceRegisteredAsSingleton() {
-    //given
-    container.register(instance: ServiceImp1() as Service)
-    
-    //when
-    let service1 = container.resolve() as Service
-    let service2 = container.resolve() as Service
-    
-    //then
-    XCTAssertTrue((service1 as! ServiceImp1) === (service2 as! ServiceImp1))
+    XCTAssertTrue(resolveDependenciesCalled)
   }
   
 }