+++ Dip/DipTests/RuntimeArgumentsTests.swift
@@ -1,9 +1,25 @@
 //
-//  RuntimeArgumentsTests.swift
-//  DipTests
+// Dip
 //
-//  Created by Ilya Puchka on 04.11.15.
-//  Copyright © 2015 AliSoftware. All rights reserved.
+// Copyright (c) 2015 Olivier Halligon <olivier@halligon.net>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 //
 
 import XCTest
@@ -34,10 +50,10 @@ class RuntimeArgumentsTests: XCTestCase {
   func testThatItResolvesInstanceWithOneArgument() {
     //given
     let arg1 = 1
-    container.register { (a1: Int) -> Service in
+    container.register(factory: { (a1: Int) -> Service in
       XCTAssertEqual(a1, arg1)
       return ServiceImp1()
-    }
+    })
     
     //when
     let service = container.resolve(arg1) as Service
@@ -150,8 +166,8 @@ class RuntimeArgumentsTests: XCTestCase {
   func testThatItRegistersDifferentFactoriesForDifferentTypesOfArguments() {
     //given
     let arg1 = 1, arg2 = "string"
-    container.register { (a1: Int) in ServiceImp1() as Service }
-    container.register { (a1: String) in ServiceImp2() as Service }
+    container.register(factory: { (a1: Int) in ServiceImp1() as Service })
+    container.register(factory: { (a1: String) in ServiceImp2() as Service })
     
     //when
     let service1 = container.resolve(arg1) as Service